+++ cmake/install.cmake
@@ -53,3 +53,12 @@ install(
     DESTINATION
         ${CMAKE_INSTALL_DOCDIR}
 )
+
+install(
+    FILES
+        ${magicxx_LICENSE_FILE}
+    DESTINATION
+        ${CMAKE_INSTALL_DOCDIR}
+    COMPONENT
+        Documentation
+)